parent and teacher view of comment

+++ Backend/remove_teachers.php
@@ -2,16 +2,36 @@
 include 'db.php';
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $id =$_POST['id'];
 
-    $sql = "DELETE FROM Teacher WHERE Teacher_ID='$id'";
+    $id =$_POST['teacherid'];
 
-    if ($conn->query($sql) === TRUE) {
-        echo "Teacher removed successfully!";
-    } else {
-        echo "Error: " . $conn->error;
+    $sql="SELECT Teacher_ID FROM Teachers";
+    $result=$conn->query($sql);
+    
+    if($result->num_rows>0){
+        while($atid= $result->fetch_assoc()){
+            if($atid==$id){
+
+                $sql = "DELETE FROM Teacher WHERE Teacher_ID='$id'";
+
+                if ($conn->query($sql) === TRUE) {
+                    echo "Teacher removed successfully!";
+                }       
+                else {
+                    echo "Error: " . $conn->error;
+                }
+                break;
+            }
+        if($atid!=$id){
+            echo "There is no teacher havind id ".$id;
+        }
+        }
+    }
+    else{
+        echo "No Teachers in the list."
     }
 
     $conn->close();
 }
+
 ?>
